2.0 Enable split build and source directories
If one is defined, use a base directory to place the build artifacts for
all project modules thus allowing the project source files and build
files to be in separate directories and possbily separate file systems,
in case that matters.  This is particularly well suited to using Google
Drive as a backup for work not yet pushed up to GitHub and to facilitate
the use of multiple development systems.