3.1 Add basic coverage tests
Add tests such that the code coverage exceeds eighty percent.  This is
done without modifying the starter app whatsoever to establish a bar
that is consistently met as the application code is modified to provide
new features.  In so doing, keeping that high coverage level is much
easier than trying to get there from a low to medium level well into the
project.